Change Category of Lists

+++ html/database/list.php
@@ -48,59 +48,64 @@
         }
     }
 
-    function addList($title, $description, $date) {
+    function addList($title, $description, $date, $category) {
         global $dbh;
         if (isset($_SESSION['username'])) {
-            $stmt = $dbh->prepare('INSERT INTO todolists (listID, username, title, descr, creation) VALUES(?, ?, ?, ?, ?)');
-            $stmt->execute(array(NULL, $_SESSION['username'],$title, $description, $date));
+            $stmt = $dbh->prepare('INSERT INTO todolists (listID, username, title, descr, creation, category) VALUES(?, ?, ?, ?, ?, ?)');
+            $stmt->execute(array(NULL, $_SESSION['username'],$title, $description, $date, $category));
         }
     }
 
 
     function addItem($id, $listID, $description, $solved, $assigneed, $datedue) {
         global $dbh;
-		$stmt = $dbh->prepare('SELECT * FROM users WHERE username = ?');
-		$stmt->execute(array($assigneed));
-		if($stmt->fetch()){
-			$stmt = $dbh->prepare('INSERT INTO listitems (id, listID, descr, solved, assignee, datedue) VALUES(?, ?, ?, ?, ?, ?)');
-			$stmt->execute(array($id, $listID, $description, $solved, $assigneed, $datedue));
-			return 0;
-		}
-		else{
-			return 1;
-		}
+        if (isset($_SESSION['username'])) {
+      		$stmt = $dbh->prepare('SELECT * FROM users WHERE username = ?');
+      		$stmt->execute(array($assigneed));
+      		if($stmt->fetch()){
+      			$stmt = $dbh->prepare('INSERT INTO listitems (id, listID, descr, solved, assignee, datedue) VALUES(?, ?, ?, ?, ?, ?)');
+      			$stmt->execute(array($id, $listID, $description, $solved, $assigneed, $datedue));
+      			return 0;
+      		}
+      		else{
+      			return 1;
+      		}
+        }
     }
-
 
     function getItemsAfterId($listID, $itemid) {
       global $dbh;
-
+      if (isset($_SESSION['username'])) {
         $stmt = $dbh->prepare('SELECT * FROM listitems WHERE id > ? AND listID = ?');
         $stmt->execute(array($itemid, $listID));
         return $stmt->fetchAll();
+      }
     }
 
 	function removeList($listID){
 		global $dbh;
+    if (isset($_SESSION['username'])) {
 
-		$stmt1 = $dbh->prepare('DELETE FROM todolists WHERE listID = ?');
-    $stmt1->execute(array($listID));
+  		$stmt1 = $dbh->prepare('DELETE FROM todolists WHERE listID = ?');
+      $stmt1->execute(array($listID));
 
-    $stmt2 = $dbh->prepare('DELETE FROM listitems WHERE listID = ?');
-    $stmt2->execute(array($listID));
-
+      $stmt2 = $dbh->prepare('DELETE FROM listitems WHERE listID = ?');
+      $stmt2->execute(array($listID));
+    }
 	}
 
     function checkItem($itemid, $listID, $solved, $username) {
       global $dbh;
+      if (isset($_SESSION['username'])) {
+
 	  /*$stmt = $dbh->prepare('SELECT * FROM todolists WHERE listID = ?');
 	  $stmt = $dbh->execute(array($listID));
 	  $listOwner = $stmt->fetchAll()['username'];
-	  
+
 	  $stmt = $dbh->prepare('SELECT * FROM listitems WHERE id = ? AND listID = ?');
 	  $stmt = $dbh->execute(array($itemid, $listID));
 	  $assignee = $stmt->fetchAll()['assignee'];
-	  
+
 	  if(($username === $assignee) || ($username === listOwner)){ */
 		$stmt = $dbh->prepare('UPDATE listitems SET solved = ? WHERE id = ? AND listID = ?');
 		$stmt->execute(array($solved, $itemid, $listID));
@@ -109,15 +114,34 @@
 	  else{
 		  return 1;
 	  }*/
-
+      }
     }
 
 
     function getListsAssociated() {
       global $dbh;
-
-        $stmt = $dbh->prepare('SELECT * FROM todolists WHERE listid = (SELECT listid FROM listitems WHERE assignee = ?) AND username != ?');
+      if (isset($_SESSION['username'])) {
+        $stmt = $dbh->prepare('SELECT * FROM todolists WHERE listID = (SELECT listID FROM listitems WHERE assignee = ?) AND username != ?');
         $stmt->execute(array($_SESSION['username'], $_SESSION['username']));
         return $stmt->fetchAll();
+      }
+    }
+
+    function getListsByCategory($category) {
+        global $dbh;
+        if (isset($_SESSION['username'])) {
+            $stmt = $dbh->prepare('SELECT * FROM todolists WHERE username = ? AND category = ?');
+            $stmt->execute(array($_SESSION['username'], $category));
+            return $stmt->fetchAll();
+        }
+    }
+
+    function getAllCategories() {
+        global $dbh;
+        if (isset($_SESSION['username'])) {
+            $stmt = $dbh->prepare('SELECT category FROM todolists WHERE username = ?');
+            $stmt->execute(array($_SESSION['username']));
+            return $stmt->fetchAll();
+        }
     }
 ?>
